src/requirements.txt
opcua
Pillow
PySimpleGUI
PyYaml
coloredlogs